Correct Javadoc and role relationships
